finish task with a avgvideosViews

+++ SocialNetworkAnalyzer/Models/TikTokUserViewModel.cs
@@ -8,7 +8,7 @@
         public string userSumLikes { get; set; }
         public string userDescription { get; set; }
         public string userImg { get; set; }
-        public double userAwgViews { get; set; }
+        public int userAwgViews { get; set; }
         public List<TikTokVideo> videos { get; set; }
     }
 }
